# docs/swagger.yaml
basePath: /api/v1
definitions:
  admin.AdminInvoiceRow:
    properties:
      created_at:
        type: string
      creator_email:
        type: string
      currency:
        type: string
      id:
        type: string
      invoice_number:
        type: string
      payer_email:
        type: string
      status:
        type: string
      total:
        type: number
    type: object
  admin.AdminStats:
    properties:
      active_users:
        type: integer
      overdue_count:
        type: integer
      paid_invoices:
        type: integer
      total_invoices:
        type: integer
      total_revenue:
        type: number
      total_users:
        type: integer
    type: object
  admin.AdminUserRow:
    properties:
      created_at:
        type: string
      email:
        type: string
      email_verified:
        type: boolean
      first_name:
        type: string
      id:
        type: string
      is_active:
        type: boolean
      last_name:
        type: string
      role:
        type: string
    type: object
...
  utils.ApiResponse:
    properties:
      data: {}
      error:
        type: string
      message:
        type: string
      status_code:
        type: integer
    type: object
host: localhost:4000
info:
  contact:
    email: davidhero125@gmail.com
    name: API Support
  description: Invoice management and payment platform API
  license:
    name: MIT
  termsOfService: http://swagger.io/terms/
  title: Stellance API
  version: "1.0"
paths:
  /admin/invoices:
    get:
      description: Paginated list of invoices with optional status and search filters
      parameters:
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Items per page
        in: query
        name: limit
        type: integer
      - description: Filter by status (draft, paid, overdue, cancelled)
        in: query
        name: status
        type: string
      - description: Search by invoice number or email
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.ApiResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/utils.ApiResponse'
      security:
      - BearerAuth: []
      summary: List all invoices
      tags:
      - admin
  /admin/invoices/{id}:
    get:
      description: Returns invoice details for admin view
      parameters:
      - description: Invoice ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.ApiResponse'
            - properties:
                data:
                  $ref: '#/definitions/admin.AdminInvoiceRow'
              type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.ApiResponse'
      security:
      - BearerAuth: []
      summary: Get invoice by ID (admin)
      tags:
      - admin
  /admin/stats:
    get:
      description: Returns total users, invoices, revenue, and overdue counts
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.ApiResponse'
            - properties:
                data:
                  $ref: '#/definitions/admin.AdminStats'
              type: object
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/utils.ApiResponse'
      security:
      - BearerAuth: []
      summary: Get platform statistics
      tags:
      - admin
  /admin/transactions:
    get:
      description: Paginated list of all platform transactions
      parameters:
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Items per page
        in: query
        name: limit
        type: integer
      - description: Search by user email or invoice number
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.ApiResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/utils.ApiResponse'
      security:
      - BearerAuth: []
      summary: List all transactions
      tags:
      - admin
  /admin/users:
    get:
      description: Paginated list of users with optional search
      parameters:
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Items per page
        in: query
        name: limit
        type: integer
      - description: Search by email or name
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.ApiResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/utils.ApiResponse'
      security:
      - BearerAuth: []
      summary: List all users
      tags:
      - admin
  /admin/users/{id}:
    delete:
      description: Permanently deletes a user account
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.ApiResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.ApiResponse'
      security:
      - BearerAuth: []
      summary: Delete user
      tags:
      - admin
    get:
      description: Returns full user details for admin view
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.ApiResponse'
            - properties:
                data:
                  $ref: '#/definitions/admin.AdminUserRow'
              type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.ApiResponse'
      security:
      - BearerAuth: []
      summary: Get user by ID
      tags:
      - admin
  /admin/users/{id}/activate:
    patch:
      description: Sets user is_active = true
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.ApiResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.ApiResponse'
      security:
      - BearerAuth: []
      summary: Activate user account
      tags:
      - admin
  /admin/users/{id}/deactivate:
    patch:
      description: Sets user is_active = false
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.ApiResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.ApiResponse'
      security:
      - BearerAuth: []
      summary: Deactivate user account
      tags:
      - admin
...
securityDefinitions:
  BearerAuth:
    description: Enter "Bearer {token}"
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"